moving renderers registry to OSS (#28986)
