+++ x-pack/plugins/canvas/public/components/element_content/index.js
@@ -4,11 +4,11 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
+import { renderFunctionsRegistry } from 'plugins/interpreter/render_functions_registry';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { compose, withProps } from 'recompose';
 import { get } from 'lodash';
-import { renderFunctionsRegistry } from '../../lib/render_functions_registry';
 import { getSelectedPage, getPageById } from '../../state/selectors/workpad';
 import { ElementContent as Component } from './element_content';
