--- x-pack/plugins/canvas/public/lib/render_functions_registry.js
@@ -1,16 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License;
- * you may not use this file except in compliance with the Elastic License.
- */
-
-import { Registry } from '@kbn/interpreter/common';
-import { RenderFunction } from './render_function';
-
-class RenderFunctionsRegistry extends Registry {
-  wrapper(obj) {
-    return new RenderFunction(obj);
-  }
-}
-
-export const renderFunctionsRegistry = new RenderFunctionsRegistry();
